requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.3
opcua==0.98.13